refactor: DaedalusConst
